Added basic visuals to game, enemies move and can damage base. Level class created, will need extra reformatting for optimization.

+++ app/src/main/java/com/example/towerdefense/MovementStrategy.java
@@ -4,6 +4,6 @@ import android.graphics.PointF;
 
 public interface MovementStrategy {
 
-    PointF move(PointF curLocation, int heading, int speed, int fps);
+    PointF move(PointF curLocation, int heading, int speed, long fps);
 
 }
